resources side menu
* resources now all in single yaml file
* yaml file generation script moved to /scripts
* side menu liquid template generates count
