--- _data/resources/assessment.yaml
@@ -1,17 +1,0 @@
-- Author: ''
-  Country: Global
-  Description: Article that investigates how to evaluate or compare open government
-    data initiatives (ODI). With initiatives like data.gov and data.gov.uk well established,
-    and well over 100 national and city-level open data initiatives emerging across
-    the globe, the question of evaluating these initiatives is coming up more and
-    more.
-  Keywords: ''
-  Language: English
-  Link: http://www.opendataimpacts.net/2011/11/evaluating-open-government-data-initiatives-can-a-5-star-framework-work/
-  MediaType: Html article
-  Notes: ''
-  Publishing date: '2011'
-  Publishing entity: Open Data Impacts
-  Region: ''
-  'Title ': 'Evaluating open government data initiatives: can a 5-star framework work?'
-  Topic: Assessment
