--- _data/resources/civic_participation.yaml
@@ -1,19 +1,0 @@
-- Author: J. Preston Whitt
-  Country: Brazil, Chile, Colombia, Dominican Republic, El Salvador, Honduras, Mexico,
-    Paraguay, Peru, Uruguay
-  Description: 'This is a report about the participation of civil society in Open
-    Gvoernment policies in 10 Latin American countries. The report is based on the
-    independant report mechanism of the Open Government Partenrship. It is also suggests
-    different ways to improve engagment with the community. '
-  Keywords: Civic participation, report mechanism, OGP
-  Language: English, Spanish
-  Link: http://www.opengovpartnership.org/sites/default/files/attachments/LatAm%20Participation%20Web.pdf
-  MediaType: PDF
-  Notes: !!python/str "http://www.opengovpartnership.org/sites/default/files/attachments/Participaci\xF3\
-    n_C%C3%ADvica_Am\xE9rica_Latina.pdfSpanish version"
-  Publishing date: '2014'
-  Publishing entity: OGP
-  Region: Latin America
-  'Title ': Independent Reporting Mechanism - Civic Participation in Latin American
-    OGP Commitments
-  Topic: Civic participation
