--- _data/resources/developing_countries.yaml
@@ -1,18 +1,0 @@
-- Author: ''
-  Country: Global
-  Description: !!python/str "Emerging Insights from Phase I of 'Exploring the Emerging\
-    \ Impacts of Open Data in Developing Countries'; a multicountry, multi-year study\
-    \ led by the World Wide Web Foundation to understand how open data is being put\
-    \ to use in different countries and contexts across the developing world. The\
-    \ project is funded by grant 107075 from Canada\u2019s International Development\
-    \ Research Center (idrc.ca) and runs from February 2013 until mid-2015."
-  Keywords: ''
-  Language: English
-  Link: http://opendataresearch.org/sites/default/files/publications/Phase%201%20-%20Synthesis%20-%20Full%20Report-print.pdf
-  MediaType: PDF
-  Notes: ''
-  Publishing date: '2014'
-  Publishing entity: Open Data Research Network
-  Region: ''
-  'Title ': 'Open Data in Developing Countries: Emerging Insights from Phase I'
-  Topic: Developing Countries
