--- _data/resources/digital_by_default.yaml
@@ -1,14 +1,0 @@
-- Author: UK Cabinet Office
-  Country: UK
-  Description: The Digital by Default Service Standard is a set of criteria for digital
-    teams building government services to mee
-  Keywords: ''
-  Language: English
-  Link: https://www.gov.uk/service-manual/digital-by-default
-  MediaType: Html article
-  Notes: ''
-  Publishing date: ''
-  Publishing entity: UK Government Digital Service
-  Region: Europe
-  'Title ': Digital by Default Service Standard
-  Topic: Digital by Default
