--- _data/resources/e-government.yaml
@@ -1,13 +1,0 @@
-- Author: Zaigham Mahmood
-  Country: ''
-  Description: ''
-  Keywords: ''
-  Language: English
-  Link: http://www.igi-global.com/book/government-implementation-practice-developing-countries/73667
-  MediaType: ''
-  Notes: ''
-  Publishing date: '2013'
-  Publishing entity: IGI Global
-  Region: ''
-  'Title ': E-Government Implementation and Practice in Developing Countries
-  Topic: E-government
