--- _data/resources/freedom_of_information.yaml
@@ -1,15 +1,0 @@
-- Author: TedxSantaCruz
-  Country: USA
-  Description: Martha Mendoza talks about the importance of open government and about
-    using Freedom of Information laws around the globe to ensure transparency. Basic
-    video to explain the concept of Open Data
-  Keywords: Transperancy, Access to information
-  Language: English
-  Link: https://www.youtube.com/watch?v=KzDE7D52zlA
-  MediaType: Video
-  Notes: ''
-  Publishing date: '2012'
-  Publishing entity: TEDx Santa Cruz
-  Region: Global
-  'Title ': Why Open Government is So Crucial To Our Society
-  Topic: Freedom of Information
