--- _data/resources/institutional_framework.yaml
@@ -1,16 +1,0 @@
-- Author: ''
-  Country: Global
-  Description: Account of a small team of independent evaluators who was invited to
-    look at the Open Data in Developing Countries project, including talking to partners
-    and collecting evidence to provide insights into the process, and the early impacts,
-    of the project.
-  Keywords: ''
-  Language: English
-  Link: http://www.opendataresearch.org/content/2014/732/guest-post-building-global-capacity-open-data
-  MediaType: Html article
-  Notes: ''
-  Publishing date: '2014'
-  Publishing entity: Open Data Research Network
-  Region: ''
-  'Title ': Building Global Capacity on Open Data
-  Topic: Institutional Framework
