--- _data/resources/no_topic.yaml
@@ -1,27 +1,0 @@
-- Author: ''
-  Country: ''
-  Description: This resource gives a short and useful explaination about what is metadata
-  Keywords: ''
-  Language: English
-  Link: http://www.theguardian.com/technology/interactive/2013/jun/12/what-is-metadata-nsa-surveillance#meta=0000000
-  MediaType: ''
-  Notes: Open is never once mentioned in the article. It's a useful exercise in metadata
-    and social media/digital tools, but otherwise not relevant to open.
-  Publishing date: '2014'
-  Publishing entity: The Guardian
-  Region: ''
-  'Title ': The Guardian Guide to Metadata
-  Topic: ''
-- Author: ''
-  Country: ''
-  Description: Connectivity Measurements
-  Keywords: ''
-  Language: ''
-  Link: ''
-  MediaType: ''
-  Notes: ''
-  Publishing date: ''
-  Publishing entity: ''
-  Region: ''
-  'Title ': ''
-  Topic: ''
