--- _data/resources/open_data_open_government.yaml
@@ -1,13 +1,0 @@
-- Author: Web Foundation
-  Country: Chile, Ghana, Indonesia
-  Description: Open Data Assessment studies from Chile, Ghana and Indonesia
-  Keywords: 'Open Data, Open Government '
-  Language: English
-  Link: http://webfoundation.org/projects/open-government-data-feasibility-studies/
-  MediaType: PDF
-  Notes: ''
-  Publishing date: 2010 - 2011
-  Publishing entity: Web Foundation
-  Region: 'Asia, LATAM, Africa '
-  'Title ': Web foundation ODRAs
-  Topic: 'Open Data, Open Government '
