--- _data/resources/open_data_principles.yaml
@@ -1,31 +1,0 @@
-- Author: UK Cabinet Office
-  Country: UK
-  Description: This document is a summary of the responses the British Cabinet Office
-    received to its public consultation, Making Open Data Real, published on 4 August
-    2011.The consultation document set out a series of questions aimed at stimulating
-    debate on how best to embed a culture of openness and transparency within public
-    services.
-  Keywords: Public consultation, Open Data, Government
-  Language: English
-  Link: https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/78885/making-data-real-consultation-summary-responses.pdf
-  MediaType: PDF
-  Notes: ''
-  Publishing date: '2011'
-  Publishing entity: UK Cabinet Office
-  Region: Europe
-  'Title ': Consultation outcome - Making open data real
-  Topic: Open Data principles
-- Author: UK Cabinet Office
-  Country: UK
-  Description: This document describe the open data principles and framework for the
-    British Government
-  Keywords: ''
-  Language: English
-  Link: https://www.gov.uk/service-manual/technology/open-data.html
-  MediaType: Html article
-  Notes: ''
-  Publishing date: '2013'
-  Publishing entity: UK Government Digital Service
-  Region: Europe
-  'Title ': British Government Open Data Prinicples
-  Topic: Open Data principles
