--- _data/resources/open_government.yaml
@@ -1,27 +1,0 @@
-- Author: Daniel Lathrop, Laurel Ruma
-  Country: 'International '
-  Description: A collection of early articles and case studies discussing various
-    aspects of open government and open data
-  Keywords: 'Open Data, Open Government '
-  Language: English
-  Link: https://github.com/oreillymedia/open_government/
-  MediaType: 'Book '
-  Notes: ''
-  Publishing date: '2010'
-  Publishing entity: O'rielly
-  Region: 'Global '
-  'Title ': Open Government
-  Topic: 'Open Government '
-- Author: OECD
-  Country: Latin America
-  Description: Case studies on Open Data in LATAM
-  Keywords: Open Government, Latin America
-  Language: English / French
-  Link: http://www.oecd-ilibrary.org/governance/open-government-policies-in-latin-america_9789264223639-en
-  MediaType: Website, PDF
-  Notes: ''
-  Publishing date: '2014'
-  Publishing entity: OECD
-  Region: Latin America
-  'Title ': Open Government in Latin America
-  Topic: Open Government
