--- _data/resources/pipeline.yaml
@@ -1,16 +1,0 @@
-- Author: ''
-  Country: Global
-  Description: Manifesto that states that we need a Data Revolution that sets a new
-    political agenda, that puts existing data to work, that improves the way data
-    is gathered and ensures that information can be used. It outlines the 12 steps
-    involved in the Data Revolution.
-  Keywords: ''
-  Language: English
-  Link: http://devinit.org/policy-briefings/data-manifesto-2/
-  MediaType: Html article
-  Notes: ''
-  Publishing date: '2014'
-  Publishing entity: Development Initiatives
-  Region: ''
-  'Title ': The Data Manifesto
-  Topic: Pipeline
