--- _data/resources/what_is_open_data.yaml
@@ -1,16 +1,0 @@
-- Author: ''
-  Country: Global
-  Description: This working paper seeks to contribute to the conversation on open
-    data research, focussing in particular on open data in developing countries. It
-    offers a brief overview of open data definitions and recent development, before
-    turning to look at different approaches for researching open data.
-  Keywords: ''
-  Language: English
-  Link: http://www.opendataresearch.org/sites/default/files/posts/Researching%20the%20emerging%20impacts%20of%20open%20data.pdf
-  MediaType: PDF
-  Notes: ''
-  Publishing date: '2013'
-  Publishing entity: Open Data Research Network
-  Region: ''
-  'Title ': Researching the emerging impacts of open data - ODDC conceptual framework
-  Topic: What is Open Data
