+++ _data/resources_filters.yaml
@@ -1,0 +1,8 @@
+---
+# This file is used to build the side menu, any item here will appear in the side menu
+# if it is a column in resources.yaml
+- Topic
+- Country
+- Region
+- Language
+- MediaType
